requirements.txt
robotframework==3.0
robotframework-selenium2library==1.7.4
robotframework-debuglibrary==0.4
Selenium==2.53.1
twine==1.6.3
zest.releaser==6.6.4
